requirements.txt
seaborn==0.12.2
paramiko==3.1.0
psycopg2-binary==2.9.6
sqlalchemy==1.4.46
